requirements.txt
easyocr==1.7.1
keras==2.13.1
nltk==3.8.1
numpy==1.23.5
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
Pillow==9.4.0
Pillow==10.3.0
regex==2023.6.3
streamlit==1.29.0
tensorflow==2.13.0
tensorflow_intel==2.13.0
transformers==4.35.0
